fix: DRY原則に従って型定義を共通化してCIエラーを修正
- coreパッケージにtypes.tsを作成して共通の型定義を集約
- AudioConfig、ConnectionConfig、FullConfigを共通化
- audioパッケージからcoreの型定義をインポート
- 型定義の重複を解消してビルドエラーを修正
